% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-matrix.R
\name{as_item_matrix}
\alias{as_item_matrix}
\title{Coerce \code{data.frame} and \code{matrix} classes to an Item Matrix.}
\usage{
as_item_matrix(x, ...)
}
\arguments{
\item{x}{Either a \code{data.frame} or \code{matrix}.}

\item{...}{Not used}
}
\value{
An \code{item_matrix} object.`
}
\description{
Converts underlying data into an Item Matrix frame.
}
\details{
\code{\link[=item_matrix]{item_matrix()}} acts as an alias for \code{\link[=as_item_matrix]{as_item_matrix()}}
}
\examples{
# Item matrix values
x = matrix(c(1, 0, 0, 1), nrow = 2)

# Construct item class
item_mat = as_item_matrix(x)
}
\seealso{
\code{\link[=item_matrix]{item_matrix()}}
}
