% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attribute_gen_bijection}
\alias{attribute_gen_bijection}
\title{Generate a vector to map polytomous vector to integers}
\usage{
attribute_gen_bijection(K, M)
}
\arguments{
\item{K}{Number of Attributes}

\item{M}{Number of Responses}
}
\value{
A \code{vec} with length \eqn{K} detailing the power's of \eqn{M}.
}
\description{
Creates a general bijection vector.
}
\details{
The bijection vector generated is
\eqn{\mathbf v = (M^{K-1},M^{K-2}, \dots, 1)^\top}.
}
\examples{
## Construct an attribute bijection for M responses ----
biject_ternary = attribute_gen_bijection(5, 3)
biject_ternary

## Construct an attribute bijection for binary responses ----
biject_binary = attribute_gen_bijection(5, 2)
biject_binary

## Construct an attribute bijection for binary responses ----
biject_default = attribute_bijection(5)
biject_default
}
\seealso{
\code{\link[=attribute_inv_gen_bijection]{attribute_inv_gen_bijection()}} and \code{\link[=attribute_bijection]{attribute_bijection()}}
}
\author{
Steven Andrew Culpepper and James Joseph Balamuta
}
