% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attribute_inv_gen_bijection}
\alias{attribute_inv_gen_bijection}
\title{Create the K inverse bijection of attribute vectors}
\usage{
attribute_inv_gen_bijection(K, M, CL)
}
\arguments{
\item{K}{Number of Attributes}

\item{M}{Number of Options.}

\item{CL}{Class Number from 0 to (2^K)-1.}
}
\value{
Return a matrix containing the class table
}
\description{
Converts the class into a bijection.
}
\examples{
## Construct an attribute bijection for M responses ----
inv_biject_ternary = attribute_inv_gen_bijection(5, 3, 2)
inv_biject_ternary

## Binary data
inv_biject_binary = attribute_inv_gen_bijection(5, 2, 4)
inv_biject_binary

## Default binary data
inv_biject_default = attribute_inv_bijection(5, 4)
inv_biject_default
}
\seealso{
\code{\link[=attribute_inv_bijection]{attribute_inv_bijection()}}, \code{\link[=attribute_gen_bijection]{attribute_gen_bijection()}}
}
\author{
James Joseph Balamuta and Steven Andrew Culpepper
}
