% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdmcore-package.R
\docType{package}
\name{ecdmcore-package}
\alias{ecdmcore}
\alias{ecdmcore-package}
\title{ecdmcore: Computational Backend for the ECDM Framework}
\description{
Provides common functions used throughout other ECDM packages.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
