% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edmcore-package.R
\docType{package}
\name{edmcore-package}
\alias{edmcore}
\alias{edmcore-package}
\title{edmcore: Algorithms used across Exploratory Diagnostic Modeling}
\description{
Provides a common set of functions used in the exploratory
    diagnostic ('EDM') framework. These functions include standard
    object definitions for Q matrices, model construction, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tmsalab.github.io/edmcore}
  \item \url{https://github.com/tmsalab/edmcore}
  \item Report bugs at \url{https://github.com/tmsalab/edmcore/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID}) [copyright holder]
  \item Jeffrey Douglas \email{jeffdoug@illinois.edu} [copyright holder]
}

}
\keyword{internal}
