% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{extract_q_matrix}
\alias{extract_q_matrix}
\alias{extract_q_matrix.q_matrix}
\alias{extract_q_matrix.default}
\title{Extract Q Matrix}
\usage{
extract_q_matrix(x, ...)

\method{extract_q_matrix}{q_matrix}(x, ...)

\method{extract_q_matrix}{default}(x, ...)
}
\arguments{
\item{x}{An `edina`, `dina`, `errum`, or `rrum` object}

\item{...}{Additional parameters}
}
\value{
A `matrix` that is either dichotomous or estimated.
}
\description{
Given a modeling object, extract the Q Matrix
}
