% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-matrix.R
\name{is_item_matrix}
\alias{is_item_matrix}
\title{Is object an Item Matrix?}
\usage{
is_item_matrix(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical vector
}
\description{
Is object an Item Matrix?
}
\examples{
x = matrix(c(1, 0, 1, 1))
is_item_matrix(item_matrix(x))
}
