% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{is_q_generic}
\alias{is_q_generic}
\title{Is the \eqn{Q} Matrix Generically Identifiable?}
\usage{
is_q_generic(x)
}
\arguments{
\item{x}{A \code{\link[=q_matrix]{q_matrix()}} or \code{\link[base:matrix]{base::matrix()}} to test.}
}
\value{
A logical vector
}
\description{
Checks if the \eqn{Q} matrix is generically identified.
}
\details{
If \eqn{\mathbf{Q}} is in the generically identifiable set \eqn{\mathcal{Q}},
then it must satisfy the following conditions:
\itemize{
\item \strong{(G1)} The true \eqn{\boldsymbol{Q}} matrix takes the form of
\eqn{\boldsymbol{Q}^\top=\left[{\boldsymbol{Q}_1,\boldsymbol{Q}_2, (\boldsymbol{Q}^\ast)^\top}\right]^\top}
after row swapping, where \eqn{\boldsymbol{Q}^\ast} is a \eqn{(J-2K)\times K}
binary matrix and \eqn{\boldsymbol{Q}_1,\boldsymbol{Q}_2 \in \mathcal{Q}_g} with
\deqn{{\mathbb {Q}}_g =\left\{ \boldsymbol{Q} \in \{0,1\}^{J\times K}:
       \boldsymbol{Q} =
       \begin{bmatrix} * &{} 1 &{} *&{} \dots &{} *&{}\dots &{} *\\
                        * &{}*&{} 1 &{} \dots &{} *&{}\dots &{} *\\
             \vdots &{} \vdots &{} &{} \ddots &{} &{} &{} \vdots \\
             * &{} * &{} *&{} \dots &{} 1&{}\dots &{} *\\
             \end{bmatrix}\right\}} where \code{*} can be either 0 or 1.
\item \strong{(G2)} For any \eqn{k = 1, 2, \ldots, K}, there exists a \eqn{j_k > 2K}, such that \eqn{q_{j_k, k} = 1}.
}
}
\examples{
## Check if Q Matrix is Generically Identified ---
# Create a generically identified Q matrix
q3_generic = rbind(diag(3),
                   c(1, 1, 0),
                   c(1, 0, 1),
                   c(0, 1, 1),
                   c(1, 1, 1))

# Check if Q matrix is generic
is_q_generic(q3_generic)

# (Extra) Check if Q matrix is strict
is_q_strict(q3_generic)
}
\references{
Gu, Yuqi, and Gongjun Xu. "Sufficient and Necessary Conditions for the Identifiability of the \eqn{Q}-matrix." arXiv preprint arXiv:1810.03819 (2018).

The function implemented is a translation of the publicly available
MATLAB code from \url{https://github.com/yuqigu/Identify_Q} into \emph{C++}.
}
\seealso{
\code{\link[=is_q_strict]{is_q_strict()}}
}
\author{
James Joseph Balamuta and Steven Andrew Culpepper
}
