% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{is_q_matrix}
\alias{is_q_matrix}
\title{Is object a Q Matrix?}
\usage{
is_q_matrix(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical vector
}
\description{
Is object a Q Matrix?
}
\examples{
x = matrix(c(1, 0, 1, 1))
is_q_matrix(q_matrix(x))
}
