% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_strict_q_identified}
\alias{is_strict_q_identified}
\title{Verify Q Matrix is Strictly Identifiable}
\usage{
is_strict_q_identified(Q)
}
\arguments{
\item{Q}{The Q matrix to be checked with dimensions \eqn{K \times J}{K x J}.}
}
\value{
A \code{bool} with value either: false or true
}
\description{
Performs a check to see if Q is strictly identified or not.
}
