% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-matrix.R
\name{item_matrix}
\alias{item_matrix}
\title{Create a Q Matrix Object}
\usage{
item_matrix(x)
}
\arguments{
\item{x}{Either a \code{data.frame} or \code{matrix}.}
}
\value{
A \code{item_matrix} object.
}
\description{
Provides a way to create an object as a \code{"item_matrix"}.
}
\examples{
# Item matrix values
x = matrix(c(1, 0, 0, 1, 1, 1), ncol = 2)

# Item matrix initialization
items = item_matrix(x)

# Data Frame encoding of Q
items_df = data.frame(
   s1 = c(1, 0, 1, 1, 0, 1),
   s2 = c(0, 1, 0, 1, 1, 0)
)

# Create a Q matrix
items = item_matrix(items_df)
}
\seealso{
\code{\link[=as_item_matrix]{as_item_matrix()}}
}
