% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-functions.R
\name{link_logit}
\alias{link_logit}
\alias{link_probit}
\title{Link Functions}
\usage{
link_logit(x)

link_probit(x)
}
\arguments{
\item{x}{A \code{matrix} or \code{vector} containing values to be transformed.}
}
\value{
A \code{matrix} or \code{vector} with values transformed under the link function.
}
\description{
Functions that rescale the data according to a given distribution.
}
\seealso{
\code{\link{theta_to_beta}}
}
