% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_bias}
\alias{metric_bias}
\title{Metric for Bias}
\usage{
metric_bias(estimate, oracle, na.rm = FALSE)
}
\arguments{
\item{estimate}{Estimated values from the model.}

\item{oracle}{Known values used to generate the model.}

\item{na.rm}{A \code{logical} indicating if missing values (including NaN) should
be removed. Default: \code{FALSE}}
}
\value{
A \code{numeric} value for each parameter comparison.
}
\description{
Computes the Bias
}
\details{
The bias measures the difference between expected value of an estimated
parameter and its true value.

The metric is computed under:
\deqn{\operatorname{Bias}(\hat \theta, \theta) = E\left[\hat\theta\right] - \theta}
}
\section{Recovery Use}{

The bias may be used to understand the difference between
\eqn{\hat\theta} matrix and the oracle \eqn{\theta} matrix.
}

\examples{
# Construct data
estimate = matrix(c(1,1,2,4,3,6), nrow = 2, ncol = 3)
truth = matrix(c(1,2,3,4,5,6), nrow = 2, ncol = 3)

# Compute the bias
metric_bias(estimate, truth)
}
\seealso{
\code{\link[base:norm]{base::norm()}}
}
