% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_element_wise}
\alias{metric_element_wise}
\title{Metric for Element-Wise Accuracy}
\usage{
metric_element_wise(estimate, oracle, na.rm = FALSE)
}
\arguments{
\item{estimate}{Estimated values from the model.}

\item{oracle}{Known values used to generate the model.}

\item{na.rm}{A \code{logical} indicating if missing values (including NaN) should
be removed. Default: \code{FALSE}}
}
\value{
A single \code{numeric} value between 0 and 1.
}
\description{
Computes the element-wise accuracy of matrices.
}
\details{
The element-wise metric is also known as accuracy or
the proportion of estimated values that are equivalent to the same elements
in the oracle value.

The metric is computed under:
\deqn{\frac{1}{JK}\sum _{j=1}^J\sum _{k=1}^K\mathcal I(\hat{\theta}_{jk}=\theta_{jk})}
}
\section{Recovery Use}{

The element-wise recovery metric is best used to understand differences
between dichotomous matrices such as the \eqn{\boldsymbol{Q}} and
\eqn{\boldsymbol{\Delta}} matrices.
}

\examples{
# Construct data
estimate = matrix(c(1,1,2,4,3,6), nrow = 2, ncol = 3)
truth = matrix(c(1,2,3,4,5,6), nrow = 2, ncol = 3)

# Compute the frobenius norm
metric_element_wise(estimate, truth)
}
\seealso{
\code{\link[base:norm]{base::norm()}}
}
