% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_frobenius_norm}
\alias{metric_frobenius_norm}
\title{Metric for Frobenius Norm}
\usage{
metric_frobenius_norm(estimate, oracle, na.rm = FALSE)
}
\arguments{
\item{estimate}{Estimated values from the model.}

\item{oracle}{Known values used to generate the model.}

\item{na.rm}{A \code{logical} indicating if missing values (including NaN) should
be removed. Default: \code{FALSE}}
}
\value{
A single \code{numeric} value.
}
\description{
Computes the Frobenius norm of matrix entries
}
\details{
The Frobenius norm is an extension of the Euclidean norm to \eqn{\mathcal{K}^{n\times n}}.

The metric is computed under:
\deqn{\|A\|_{\rm F} = \left(\sum_{i=1}^m \sum_{j=1}^n |a_{ij}|^2\right)^{\frac{1}{2}}}
}
\section{Recovery Use}{

The Frobenius norm is best used to understand differences between
the estimated \eqn{\hat\theta} matrix and the oracle \eqn{\theta} matrix.
}

\examples{
# Construct data
estimate = matrix(c(1,1,2,4,3,6), nrow = 2, ncol = 3)
truth = matrix(c(1,2,3,4,5,6), nrow = 2, ncol = 3)

# Compute the frobenius norm
metric_frobenius_norm(estimate, truth)
}
\seealso{
\code{\link[base:norm]{base::norm()}}
}
