% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_matrix_wise}
\alias{metric_matrix_wise}
\title{Metric for Matrix-Wise Accuracy}
\usage{
metric_matrix_wise(estimate, oracle, na.rm = FALSE)
}
\arguments{
\item{estimate}{Estimated values from the model.}

\item{oracle}{Known values used to generate the model.}

\item{na.rm}{A \code{logical} indicating if missing values (including NaN) should
be removed. Default: \code{FALSE}}
}
\value{
A single \code{numeric} value between 0 and 1.
}
\description{
Computes the matrix-wise accuracy.
}
\details{
The matrix-wise metric is a variant of accuracy that holistically looks
at the entire estimated matrix against the entire

The metric is computed under:
\deqn{I(\hat{\theta}=\theta)}
}
\section{Recovery Use}{

The element-wise recovery metric is best used to understand differences
between dichotomous matrices such as the \eqn{\boldsymbol{Q}} and
\eqn{\boldsymbol{\Delta}} matrices.
}

\examples{
# Construct data
estimate = matrix(c(1,1,2,4,3,6), nrow = 2, ncol = 3)
truth = matrix(c(1,2,3,4,5,6), nrow = 2, ncol = 3)

# Compute the frobenius norm
metric_matrix_wise(estimate, truth)
}
\seealso{
\code{\link[base:norm]{base::norm()}}
}
