% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_mode}
\alias{metric_mode}
\title{Metric for determining the most popular value}
\usage{
metric_mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{vector} of data.}

\item{na.rm}{A \code{logical} indicating if missing values (including NaN) should
be removed. Default: \code{FALSE}}
}
\value{
A single \code{numeric} or \code{integer} value.
}
\description{
Computes the mode given a vector of data.
}
\details{
Be forewarned that this method of obtaining the mode does not take into
consideration ties. That is, only the first group with the maximum value
is returned. If a second group also has a similar count, this group will
not be returned.
}
\section{Recovery Use}{

The mode should be used when computing latent class recovery among subjects.
}

\examples{
# Sample class-data
x = c(0, 0, 0, 1, 1, 2)
metric_mode(x)
}
