% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edm-model.R
\name{new_edm_default_property_list}
\alias{new_edm_default_property_list}
\title{Construct a property list for the model.}
\usage{
new_edm_default_property_list(n, j, k, runtime, ..., class = character())
}
\arguments{
\item{n}{Number of Observations}

\item{j}{Number of Items}

\item{k}{Number of Attributes}

\item{runtime}{Duration of C++ Estimation}

\item{...}{Additional properties}

\item{class}{Designate the model-specific class.}
}
\description{
Holds detail information about the model.
}
