% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edm-model.R
\name{new_edm_model}
\alias{new_edm_model}
\title{Construct an EDM model class object with pre-defined names.}
\usage{
new_edm_model(chain, property_list, estimate = NA, ..., class = character())
}
\arguments{
\item{chain}{Values from the MCMC chains across iterations.}

\item{estimate}{Any estimate done within C++ and raised into R during the
modeling procedure.}

\item{...}{Additional model-specific data}

\item{property}{Information on the data and algorithm configuration.}
}
\value{
An EDM model object with subclasses of \code{class}, \code{edm}, and \code{list} that
contains:
\itemize{
\item \code{chains}: Chain values
\item \code{property}: Estimation details including:
\itemize{
\item \code{n}: Number of Observations
\item \code{k}: Number of Attributes
\item \code{order}: Highest degree from 1 to P = 2^K or less.
}
\item \code{estimates}:
\itemize{
\item \strong{Note:} May be \code{NA} or a sub-list.
}
}
}
\description{
Base class that provides slots for growth within the modeling objects.
}
