% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edm-summary.R
\name{new_edm_summary}
\alias{new_edm_summary}
\title{Construct an EDM Summary class object}
\usage{
new_edm_summary(property, estimate, ..., class = character())
}
\arguments{
\item{property}{Information on the data and algorithm configuration.}

\item{estimate}{Mean, SD, Quantiles on chains pased in.}

\item{...}{Additional model-specific data}
}
\value{
An EDM summary object with subclasses of \code{class}, \code{edm_summary}, and \code{list} that
contains:
\itemize{
\item \code{property}: Estimation details including:
\itemize{
\item \code{n}: Number of Observations
\item \code{k}: Number of Attributes
\item \code{order}: Highest degree from 1 to P = 2^K or less.
}
\item \code{estimate}: Mean, SD, Quantiles on MCMC chains.
}
}
\description{
Base class that provides slots for growth within the modeling objects.
}
