% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutate_attribute_level_table}
\alias{permutate_attribute_level_table}
\title{Permutation Attribute Table}
\usage{
permutate_attribute_level_table(k, m = 2)
}
\arguments{
\item{k}{Supply the number of attributes to change}

\item{m}{Number of Responses. Default: 2.}
}
\description{
Generate a table containing attribute permutations.
}
\examples{

# Example of generating all attribute level swap permutations
k = 3
nClass = 2 ^ k

# Total number of label swaps for unstructured mixture
factorial(nClass)

# Total number of label swaps for structured mixture
factorial(k) * (2 ^ k)

# Create the permutation table
permutation_table = permutate_attribute_level_table(k = 3, m = 2)

# Loop over columns to find the positions of the equivalent attribute level swaps
}
