% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutate_binary_matrix}
\alias{permutate_binary_matrix}
\title{Reorder an object to match a target as closely as possible}
\usage{
permutate_binary_matrix(current, target)
}
\arguments{
\item{current}{Present \code{matrix} object.}

\item{target}{Desired \code{matrix} object.}
}
\value{
A \code{permutated_matrix} object that contains:
\itemize{
\item \code{permutate_matrix}:
The \code{current} matrix permutated to match the \code{target}
\item \code{permutate_order}:
The permutation order applied to the columns of \code{current}.
\item \code{permutate_mean}:
The mean difference between the permutation current and target.
Bound between 0 and 1 with a mean value close to 1 indicating an
exact permutation between \code{current} and \code{target} is possible.
\item \code{permutate_target}:
Reference object used for permutation.
}
}
\description{
Rearranges columns within a matrix until the closest permutation
to the target is found.
}
