% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutate_by_matrix}
\alias{permutate_by_matrix}
\title{Label switching after the fact}
\usage{
permutate_by_matrix(estimate, oracle = NULL, M = 2)
}
\arguments{
\item{estimate}{A matrix}

\item{oracle}{A matrix. Default a random uniform matrix.}
}
\description{
Default is Check matrix
}
\references{
Erosheva, E. A., & Curtis, S. M. (2011).
Dealing with rotational invariance in Bayesian confirmatory factor analysis.
Department of Statistics, University of Washington, Seattle, Washington, USA.
\url{https://stat.uw.edu/research/tech-reports/dealing-rotational-invariance-bayesian-confirmatory-factor-analysis}
}
