% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutate_theta_order}
\alias{permutate_theta_order}
\title{Reorganize parameters in Theta Matrix}
\usage{
permutate_theta_order(estimated_theta, oracle_theta, k = 3)
}
\arguments{
\item{estimated_theta}{Estimated Theta Matrix}

\item{oracle_theta}{Known Theta Matrix}

\item{k}{Number of Attributes}
}
\value{
A vector containing the permutation order
}
\description{
Reorganize parameters in Theta Matrix
}
