% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permuteAtableIndices}
\alias{permuteAtableIndices}
\title{Permute indices}
\usage{
permuteAtableIndices(nClass, K, order, vv, perm)
}
\arguments{
\item{nClass}{Number of classes given by \eqn{2^K}.}

\item{K}{Number of Attributes}

\item{order}{Degree of Order}

\item{vv}{Bijection vector}

\item{perm}{Permutations}
}
\value{
Permuted table at specified indices.
}
\description{
Constructs and permutes indices within a vector
}
