% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess_attribute_correlation}
\alias{postprocess_attribute_correlation}
\title{Correlation Among Attribute Membership}
\usage{
postprocess_attribute_correlation(pi_hat, k, order = k)
}
\arguments{
\item{pi_hat}{Estimated latent class membership probabilities}

\item{k}{Number of attributes}

\item{order}{Number of interactions in the model}
}
\description{
Given the estimated latent class membership probabilities, we compute the
correlation across classes.
}
\examples{

# Sample calculation when K = 3 and number of classses is 2^3 = 8
pi_hat = c(0.10, 0.10, 0.01, 0.13, 0.03, 0.26, 0.05, 0.31)
k = 3

# Compute attribute correlation matrix
postprocess_attribute_correlation(pi_hat, k)
}
