% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edm-model.R
\name{print.edm_model}
\alias{print.edm_model}
\title{Print Exploratory Diagnostic Model Components}
\usage{
\method{print}{edm_model}(x, ...)
}
\arguments{
\item{x}{an edm modeling object}

\item{...}{additional arguments passed to or from other methods.}
}
\description{
Custom print method for viewing Exploratory Diagnostic Model components.
}
