% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edm-model.R
\name{print.edm_property_list}
\alias{print.edm_property_list}
\title{Display Exploratory Diagnostic Model Properties}
\usage{
\method{print}{edm_property_list}(x, ...)
}
\arguments{
\item{x}{an edm modeling object}

\item{...}{additional arguments passed to or from other methods.}
}
\description{
Custom print method for viewing Exploratory Diagnostic Model properties.
}
