% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-matrix.R
\name{print.item_matrix}
\alias{print.item_matrix}
\title{Output Item Matrix}
\usage{
\method{print}{item_matrix}(x, ...)
}
\arguments{
\item{x}{An \code{item_matrix} object}

\item{...}{Additional methods passed onto the \code{print.matrix} method.}
}
\value{
An invisible \code{matrix} without the \code{item_matrix} class displayed as a part
of the output displayed.
}
\description{
Custom print method for the Item Matrix Object.
}
\examples{
# Item matrix values
x = matrix(c(1, 0, 0, 1), nrow = 2)

# Show item matrix structure
item_matrix(x)
}
\seealso{
\code{\link[=item_matrix]{item_matrix()}}, \code{\link[=as_item_matrix]{as_item_matrix()}}
}
