% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{print.permutation_matrix}
\alias{print.permutation_matrix}
\title{Custom print handle for displaying the permutation ordering}
\usage{
\method{print}{permutation_matrix}(x, ...)
}
\arguments{
\item{x}{A \code{permutation_matrix} object.}

\item{...}{Not used.}
}
\description{
Custom print handle for displaying the permutation ordering
}
