% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{print.q_matrix}
\alias{print.q_matrix}
\title{Printing out a Q Matrix Object}
\usage{
\method{print}{q_matrix}(x, ...)
}
\arguments{
\item{x}{An `q_matrix` object}

\item{...}{Additional methods passed onto the `print.matrix` method.}
}
\description{
Custom print method for the Q Matrix Object.
}
