% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-q-matrix.R
\name{plot.q_matrix}
\alias{plot.q_matrix}
\title{Heatmap Visualization of a Q Matrix}
\usage{
\method{plot}{q_matrix}(x, ...)
}
\arguments{
\item{x}{Either an \code{edina}, \code{errum}, or \code{q_matrix} object.}

\item{...}{Additional paramters not used}
}
\value{
A \code{ggplot2} object that can be further manipulated.
}
\description{
Provides a heatmap approach to showing the estimated binary or averaged
values of the Q Matrix.
}
\examples{
q = q_matrix(matrix(c(1, 0, 1, 1, 0, 1), ncol = 3))
plot(q)
}
