% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{q_matrix}
\alias{q_matrix}
\title{Create a Q Matrix Object}
\usage{
q_matrix(x)
}
\arguments{
\item{x}{Either a \code{data.frame} or \code{matrix}.}
}
\value{
A \code{q_matrix} object.
}
\description{
Provides a way to create an object as a \code{"q_matrix"}.
}
\examples{
# Create a matrix with only 1s and 0s.
x = matrix(c(1, 0, 0, 1), ncol = 2)

# Register it as a Q matrix.
q_matrix(x)
}
