% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-format.R
\name{read_q_matrix}
\alias{read_q_matrix}
\title{Import a Q Matrix}
\usage{
read_q_matrix(file, header = FALSE, sep = " ", skip = 0)
}
\arguments{
\item{file}{name, url, or \link{connections} of the file to be read}

\item{header}{a logical value to indicate if the data contains a
description field as the first row. Default is \code{FALSE}.}

\item{sep}{the field separator character. Values on each line of the
file are separated by this character. Default is white
space given by \code{sep = " "}.}

\item{skip}{the number of lines of the data file that should be
skipped before the data is read. Default is 0.}
}
\value{
A \code{matrix} labeled with row names as \code{ItemYY} and column names as \code{SkillZZ}
}
\description{
Allows for a dichotomous Q Matrix to be imported with standard styling.
}
\details{
This function is designed to specifically read in dichotomous Q matrix
into \emph{R}. The matrix must be structured with the appropriate separator.
}
\seealso{
\code{\link[=q_matrix]{q_matrix()}} or \code{\link[=as_q_matrix]{as_q_matrix()}} for constructing a Q matrix from an
existing class.
}
