% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{recovery_element}
\alias{recovery_element}
\title{Element-wise Accuracy Table for Q Matrix Estimation}
\usage{
recovery_element(estimated, oracle, decision = 0.5)
}
\arguments{
\item{estimated}{Set containing the estimated strategies values
on generated data. Dimensions: \eqn{J x K x S}.}

\item{oracle}{Set containing the actual strategies used to
generate the data. Dimensions: \eqn{J x K x S}.}

\item{decision}{Threshold for applying a classification to estimated
entry. Default: \code{0.5}}
}
\value{
A \code{list} containing:
\itemize{
\item \code{element_wise_recovery}:
Average number of entries matching between the Q matrix and estimated
matrix.
\item \code{permutated_q}:
Ordered Q matrix by permutation under decision rule of 0.5.
\item \code{decision_rule}:
Threshold cut-off applied to the estimated Q matrix to ensure dichotomous
entries.
}
}
\description{
Given the oracle Q matrix, compare the estimated Q matrix to
determine how accurate the estimation routine is.
}
\examples{
# Create a Q matrix
Q_oracle =
  rbind(c(0, 1),
        c(1, 0),
        c(0, 1),
        c(1, 0),
        c(0, 1),
        c(0, 1),
        c(1, 0))

# Simulate a random Q matrix
Q_est = matrix(runif(7*2), ncol = 2)

# Obtain the recovery metric for elementwise matches
recovery_element(Q_est, Q_oracle)
}
