% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-functions.R
\name{theta_to_beta}
\alias{theta_to_beta}
\alias{theta_probit_to_beta}
\alias{theta_logit_to_beta}
\title{Translate Theta values into Beta values}
\usage{
theta_to_beta(theta, k, link_function)

theta_probit_to_beta(theta, k)

theta_logit_to_beta(theta, k)
}
\arguments{
\item{theta}{Matrix of theta values}

\item{k}{Number of Attributes}

\item{link_function}{Link function to use}
}
\value{
Beta matrix after translation.
}
\description{
Converts theta values underneath a link function to beta values.
}
\seealso{
\code{\link{link_logit}} and \code{\link{link_probit}}
}
