% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attribute_gen_bijection}
\alias{attribute_gen_bijection}
\title{Generate a vector to map polytomous vector to integers}
\usage{
attribute_gen_bijection(K, M)
}
\arguments{
\item{K}{Number of Attributes}

\item{M}{Number of Responses}
}
\value{
Return a matrix containing the class table
}
\description{
Converts class into a bijection to integers
}
