% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attribute_inv_gen_bijection}
\alias{attribute_inv_gen_bijection}
\title{Create the K inverse bijection of attribute vectors}
\usage{
attribute_inv_gen_bijection(K, M, CL)
}
\arguments{
\item{K}{Number of Attributes}

\item{M}{Number of Options.}

\item{CL}{Class Number from 0 to (2^K)-1.}
}
\value{
Return a matrix containing the class table
}
\description{
Converts the class into a bijection.
}
