% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdmcore-package.R
\docType{package}
\name{ecdmcore-package}
\alias{ecdmcore}
\alias{ecdmcore-package}
\title{ecdmcore: Algorithms used across Exploratory Cognitive Diagnostic Modeling}
\description{
Provides a common set of functions used in the exploratory
    cognitive diagnostic ('ECDM') framework. These functions include standard
    object definitions for Q matrices, model construction, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tmsalab.github.io/ecdmcore}
  \item \url{https://github.com/tmsalab/ecdmcore}
  \item Report bugs at \url{https://github.com/tmsalab/ecdmcore/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (0000-0003-4226-6176) [copyright holder]
  \item Jeffrey Douglas \email{jeffdoug@illinois.edu} [copyright holder]
}

}
\keyword{internal}
