% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-matrix.R
\name{is_item_matrix}
\alias{is_item_matrix}
\title{Is object an Item Matrix?}
\usage{
is_item_matrix(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical vector
}
\description{
Is object an Item Matrix?
}
\examples{
# Specify int matrix details
x = matrix(c(1, 0, 1, 1, 1), ncol = 2)

# Coerce to an item matrix
items = item_matrix(x)

# Verify output is an item matrix
is_item_matrix(items)

# Not an item matrix, but a regular matrix...
is_item_matrix(x)
}
