% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{is_q_strict}
\alias{is_q_strict}
\title{Is Q Matrix Strictly or Generically Identifiable?}
\usage{
is_q_strict(x)
}
\arguments{
\item{x}{A \code{\link[=q_matrix]{q_matrix()}} or \code{\link[base:matrix]{base::matrix()}} to test.}
}
\value{
A logical vector
}
\description{
Is Q Matrix Strictly or Generically Identifiable?
}
\section{Strictly Identifiable}{


Add details...
}

\section{Generically Identifiable}{


Add details...
}

\examples{
x = matrix(c(1, 0, 1, 1))
is_q_strict(x)
is_q_strict(q_matrix(x))
}
