% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutate_attribute_level_table}
\alias{permutate_attribute_level_table}
\title{Permutation Attribute Table}
\usage{
permutate_attribute_level_table(K, M = 2)
}
\arguments{
\item{M}{Number of Responses. Default: 2.}

\item{Supply}{the number of attributes to change}
}
\description{
Generate a table containing attribute permutations.
}
\examples{

# Example of generating all attribute level swap permutations
K = 3
nClass = 2 ^ K

# Total number of label swaps for unstructured mixture
factorial(nClass)

# Total number of label swaps for structured mixture
factorial(K) * (2 ^ K)

# Create the permutation table
permutation_table = permutate_attribute_level_table(K = 3, M = 2)

# Loop over columns to find the positions of the equivalent attribute level swaps
}
