% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-recovery.R
\name{permutate_order}
\alias{permutate_order}
\title{Reorganize parameters in Theta Matrix}
\usage{
permutate_order(estimated_theta, oracle_theta, K = 3)
}
\arguments{
\item{estimated_theta}{Estimated Theta Matrix}

\item{oracle_theta}{Known Theta Matrix}

\item{K}{Number of Attributes}
}
\value{
A vector containing the permutation order
}
\description{
Reorganize parameters in Theta Matrix
}
