% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{q_matrix}
\alias{q_matrix}
\title{Create a Q Matrix Object}
\usage{
q_matrix(x)
}
\arguments{
\item{x}{Either a `data.frame` or `matrix`.}
}
\value{
A `q_matrix` object.
}
\description{
Provides a way to create an object as a `"q_matrix"`.
}
\examples{

}
